Implement driver for ueye camera UI-3251LE-M-GL
This driver is specifically tailored to work with the ueye camera UI-3251LE-M-GL,
and a fish-eye lens model Lensagon BF2M15520.

The driver creates a ROI for the area occupied for the lens, and allows
the configuration of the following parameters:

+pixel clock
+frame rate
+exposure
+black level offset
+gamma
+hardware gamma
+master gain

It publishes the video streamed by the camera to /camera/image_raw. It
also reads the calibration file but image_proc doesn't seem to be able
to rectify the images.